+++ christmas_app/static/index.js
@@ -6,7 +6,7 @@ function hit() {
     title: "Surprise Present is ready for you!",
     html: "Aren't you excited?",
     icon: "info",
-    confirmButtonText: '<i class="fa fa-thumbs-up"></i> Exciting!',
+    confirmButtonText: "Exciting!",
     confirmButtonAriaLabel: "Thumbs up, sure!",
   }).then(() => {
     window.location.assign("/login");
@@ -18,7 +18,7 @@ async function theChecker(loc = "") {
   if (loc == "login-integrated") {
     var htmlCode =
       '<form class="form" action="/check/" method="POST"> ' +
-      "<h3>Check pre-defined account</h3>" +
+      "<h3 style='font-weight: bold; color: red'>Check pre-defined account</h3>" +
       '<input class="col-8 my-2" type="search" placeholder="Search Here" aria-label="Search" autocapitalize="characters" name="account_" id="account_"/><br>' +
       // '<p>(Ex. If your name is John Blue, type in "John" or "JOHN" or "john")</p>' +
       '<input type="submit" value="Search Now!" class="btn-lg" />' +
@@ -26,7 +26,7 @@ async function theChecker(loc = "") {
   } else {
     var htmlCode =
       '<form class="form" action="/check-account/by-first-name/" method="POST"> ' +
-      "<h3>Check pre-defined account</h3>" +
+      "<h3 style='font-weight: bold; color: red'>Check pre-defined account</h3>" +
       '<input class="col-8 my-2" type="search" placeholder="Search Here" aria-label="Search" autocapitalize="characters" name="account_" id="account_"/><br>' +
       '<input type="submit" value="Search Now!" class="btn-lg" />' +
       "</form>";
@@ -111,5 +111,27 @@ async function riskLevelChecker() {
     showConfirmButton: false,
     allowOutsideClick: true,
     allowEscapeKey: true,
+  });
+}
+
+// Pop up contact info
+async function contactMe() {
+  htmlCode =
+    '<div class="phone py-1 mx-1"><strong>Tel.:</strong> <i>+66 942536546</i></div>' +
+    '<div class="email py-1 mx-1"><strong>Email:</strong> <i>kittipich.a@gmail.com</i></div>' +
+    '<div class="facetime py-1 mx-1"><strong>Facetime:</strong> <i>[through my number or email]</i></div>' +
+    '<div class="wechat py-1 mx-1"><strong>WeChat ID:</strong> <i>LukeAtMyEyes2</i></div>' +
+    '<div class="line py-1 mx-1"><strong>Line ID:</strong> <i>kan_kittipich</i></div>';
+  Swal.fire({
+    title: "Reach me via...",
+    icon: "info",
+    html: htmlCode,
+    showCancelButton: false,
+    showConfirmButton: true,
+    showLoaderOnConfirm: true,
+    confirmButtonText: "OKAY!",
+    confirmButtonAriaLabel: "okay!",
+
+    allowOutsideClick: () => !Swal.isLoading(),
   });
 }